+++ src/app/app.component.html
@@ -1,2 +1,1 @@
-<!-- <app-base-graph></app-base-graph> -->
 <app-treemap></app-treemap>
