--- src/app/data.config.ts
@@ -1,4 +1,0 @@
-export const BASE_PATH = 'assets/data/';
-export const BUS = 'bus-1062.csv';
-export const BRANCH = 'branch-1062.csv';
-
