--- src/app/layout.model.ts
@@ -1,10 +1,0 @@
-export interface ILayoutInfo { 
-    margin: {
-        top: number;
-        right: number;
-        left: number;
-        bottom: number;
-    }
-    height: number;
-    width: number;
-}
