+++ src/app/treemap/treemap.component.css
@@ -1,13 +1,18 @@
 .treemap {
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
-}
-
-
-#card {
-  padding: auto;
-  align-items: center;
-  /* color: #434343; */
-  background-color:beige;
-}
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+  }
+  
+  
+  #card {
+    padding: auto;
+    align-items: center;
+    /* color: #434343; */
+    background-color:beige;
+  }
+  
+  .apply {
+    font-size: 20px;
+    height: 48px;
+  }
