+++ src/app/treemap/treemap.component.html
@@ -1,147 +1,107 @@
 <div class="treemap">
 
-  <div>
-    <svg #rootSvg></svg>
+    <div>
+      <svg #rootSvg></svg>
+    </div>
+    <div #tooltip></div>
+  
+    <span class="metric">
+      
+  
+      <mat-form-field appearance="fill">
+        <mat-label>iter number</mat-label>
+        <input matInput type="number" [(ngModel)]="iter">
+        <button *ngIf="iter" matSuffix mat-icon-button aria-label="Clear" (click)="iter=0">
+          <mat-icon>close</mat-icon>
+        </button>
+      </mat-form-field>
+  
+      <p>time : {{bind_time}} (sec)</p>
+  
+      <p>
+        <mat-button-toggle-group [(ngModel)]="layout">
+          <mat-button-toggle value="Z_Layout">Z_Layout</mat-button-toggle>
+          <mat-button-toggle value="Sequence">Sequence</mat-button-toggle>
+          <mat-button-toggle value="Local_Random">Local_Random</mat-button-toggle>
+          <mat-button-toggle value="Global_Random">Global_Random</mat-button-toggle>
+        </mat-button-toggle-group>
+      </p>
+      
+  
+      <mat-form-field appearance="fill">
+        <mat-label>Select Data File</mat-label>
+        <mat-select [(value)]="load_data_num">
+          <mat-option *ngFor="let d of load_datas" [value]="d" >
+            {{d}}
+          </mat-option>
+        </mat-select>
+      </mat-form-field>
+  
+      <mat-form-field appearance="fill">
+        <mat-label>Select Port number</mat-label>
+        <mat-select [(value)]="port">
+          <mat-option *ngFor="let d of ports" [value]="d" >
+            {{d}}
+          </mat-option>
+        </mat-select>
+      </mat-form-field>
+  
+      <table border="1" width ="300" cellspacing="0">
+        <tr>
+          <td>Total Length</td>
+          <td>{{measurement[0]}}</td>
+        </tr>
+        <tr>
+          <td>Edge Crossing</td>
+          <td>{{measurement[1]}}</td>
+        </tr>
+        <tr>
+          <td>Total Bending</td>
+          <td>{{measurement[2]}}</td>
+        </tr>
+      </table>
+  
+      <p></p>
+      <hr/>
+      <p></p>
+      <h1>Random_statistics</h1>
+      <p>
+        <mat-button-toggle-group (valueChange)="StatisticsSelect($event)">
+          <mat-button-toggle value="Total_Length">Total Length</mat-button-toggle>
+          <mat-button-toggle value="Edge_Crossing">Edge Crossing</mat-button-toggle>
+          <mat-button-toggle value="Total_Bending">Total Bending</mat-button-toggle>
+        </mat-button-toggle-group>
+      </p>
+  
+      <table border="1" width ="300" cellspacing="0">
+        <tr>
+          <td>Mean</td>
+          <td width ="80%">{{statisticdata.mean[statistics_index]}}</td>
+        </tr>
+        <tr>
+          <td>Median</td>
+          <td width ="80%">{{statisticdata.median[statistics_index]}}</td>
+        </tr>
+        <tr>
+          <td>Min</td>
+          <td width ="80%">{{statisticdata.min[statistics_index]}}</td>
+        </tr>
+        <tr>
+          <td>Max</td>
+          <td width ="80%">{{statisticdata.max[statistics_index]}}</td>
+        </tr>
+        <tr>
+          <td>std</td>
+          <td width ="80%">{{statisticdata.std[statistics_index]}}</td>
+        </tr>
+      </table>
+  
+      <p></p>
+      <hr/>
+      <p></p>
+
+      <button class="apply" mat-raised-button (click)="apply()">apply</button>
+  
+    </span>
   </div>
-  <div #tooltip></div>
-
-  <span class="metric">
-
-    <mat-form-field appearance="fill">
-      <mat-label>iter number</mat-label>
-      <input matInput type="text" [(ngModel)]="iter">
-      <button *ngIf="iter" matSuffix mat-icon-button aria-label="Clear" (click)="iter=''">
-        <mat-icon>close</mat-icon>
-      </button>
-    </mat-form-field>
-
-    <p>time : {{bind_time}} (sec)</p>
-
-    <p>
-      <mat-button-toggle-group (valueChange)="ToggleSelect($event)">
-        <mat-button-toggle value="Z_Layout">Z_Layout</mat-button-toggle>
-        <mat-button-toggle value="Sequence">Sequence</mat-button-toggle>
-        <mat-button-toggle value="Local_Random">Local_Random</mat-button-toggle>
-        <mat-button-toggle value="Global_Random">Globla_Landom</mat-button-toggle>
-      </mat-button-toggle-group>
-    </p>
-    
-
-    <mat-form-field appearance="fill">
-      <mat-label>Select Data File</mat-label>
-      <mat-select>
-        <mat-option *ngFor="let d of data" [value]="d" (blur)="select(d)">
-          {{d}}
-        </mat-option>
-      </mat-select>
-    </mat-form-field>
-
-    <mat-form-field appearance="fill">
-      <mat-label>Select Port number</mat-label>
-      <mat-select>
-        <mat-option *ngFor="let d of port" [value]="d" (blur)="port_select(d)">
-          {{d}}
-        </mat-option>
-      </mat-select>
-    </mat-form-field>
-
-    <!-- <div class="card-list">
-      <mat-card id="card">
-        <mat-card-title>Total Length</mat-card-title>
-        <mat-card-content>{{measurement[0]}}</mat-card-content>
-      </mat-card>
-      <mat-card id="card">
-        <mat-card-title>Edge Crossing</mat-card-title>
-        <mat-card-content>{{measurement[1]}}</mat-card-content>
-      </mat-card>
-      <mat-card id="card">
-        <mat-card-title>Total Bending</mat-card-title>
-        <mat-card-content>{{measurement[2]}}</mat-card-content>
-      </mat-card>
-      <br>
-    </div> -->
-
-    <!-- <p>Total Length : {{measurement[0]}}</p>
-    <p>Edge Crossing : {{measurement[1]}}</p>
-    <p>Total Bending : {{measurement[2]}}</p> -->
-
-    <table border="1" width ="300" cellspacing="0">
-      <tr>
-        <td>Total Length</td>
-        <td>{{measurement[0]}}</td>
-      </tr>
-      <tr>
-        <td>Edge Crossing</td>
-        <td>{{measurement[1]}}</td>
-      </tr>
-      <tr>
-        <td>Total Bending</td>
-        <td>{{measurement[2]}}</td>
-      </tr>
-    </table>
-
-    <p></p>
-    <hr/>
-    <p></p>
-    <h1>Random_statistics</h1>
-    <p>
-      <mat-button-toggle-group (valueChange)="StatisticsSelect($event)">
-        <mat-button-toggle value="Total_Length">Total Length</mat-button-toggle>
-        <mat-button-toggle value="Edge_Crossing">Edge Crossing</mat-button-toggle>
-        <mat-button-toggle value="Total_Bending">Total Bending</mat-button-toggle>
-      </mat-button-toggle-group>
-    </p>
-
-    <!-- <div class="card-list">
-      <mat-card id="card">
-        <mat-card-title>Mean</mat-card-title>
-        <mat-card-content>{{random_mean[statistics_index]}}</mat-card-content>
-      </mat-card>
-      <mat-card id="card">
-        <mat-card-title>Median</mat-card-title>
-        <mat-card-content>{{random_median[statistics_index]}}</mat-card-content>
-      </mat-card>
-      <mat-card id="card">
-        <mat-card-title>Min</mat-card-title>
-        <mat-card-content>{{random_min[statistics_index]}}</mat-card-content>
-      </mat-card>
-      <mat-card id="card">
-        <mat-card-title>Max</mat-card-title>
-        <mat-card-content>{{random_max[statistics_index]}}</mat-card-content>
-      </mat-card>
-      <mat-card id="card">
-        <mat-card-title>Standard_deviation</mat-card-title>
-        <mat-card-content>{{random_std[statistics_index]}}</mat-card-content>
-      </mat-card>
-      <br>
-    </div> -->
-
-    <table border="1" width ="300" cellspacing="0">
-      <tr>
-        <td>Mean</td>
-        <td width ="80%">{{statisticdata.mean[statistics_index]}}</td>
-      </tr>
-      <tr>
-        <td>Median</td>
-        <td width ="80%">{{statisticdata.median[statistics_index]}}</td>
-      </tr>
-      <tr>
-        <td>Min</td>
-        <td width ="80%">{{statisticdata.min[statistics_index]}}</td>
-      </tr>
-      <tr>
-        <td>Max</td>
-        <td width ="80%">{{statisticdata.max[statistics_index]}}</td>
-      </tr>
-      <tr>
-        <td>std</td>
-        <td width ="80%">{{statisticdata.std[statistics_index]}}</td>
-      </tr>
-    </table>
-
-    <p></p>
-    <hr/>
-    <p></p>
-
-  </span>
-</div>
+
