--- src/app/workspace/workspace.component.html
@@ -1,1 +1,0 @@
-<p>workspace works!</p>
