--- src/app/workspace/workspace.component.ts
@@ -1,14 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-workspace',
-  templateUrl: './workspace.component.html',
-  styleUrls: ['./workspace.component.css']
-})
-export class WorkspaceComponent implements OnInit {
-
-
-  ngOnInit(): void {
-  }
-
-}
