+++ src/shared/modules/calculate edge crossing/edge-measurement.ts
@@ -1,9 +1,7 @@
 import { IBranchData } from "src/shared/interfaces/ibranch-data";
 import { IBusObjectData } from "src/shared/interfaces/ibus-object-data";
-import { seqeunce_TreemapData } from "../datas/seqeunce-treemap-data";
-import { TreemapData } from "../datas/treemap-data";
-import { local_Random_TreemapData } from 'src/shared/modules/treemap-vis/datas/local-random-treemap-data';
-import { global_Random_TreemapData } from 'src/shared/modules/treemap-vis/datas/global-random-treemap-data';
+import { TreemapNode } from "../node-placement/treemap-node.service";
+
 
 class EdgeInfo {
     public eType: number[] = [-1, -1, -1, -1]; //h=1,w=2. -1은 없음(NA)
@@ -31,7 +29,7 @@ class EdgeInfo {
 }
 
 export class EdgeMeasurement {
-    private treemapData: TreemapData|seqeunce_TreemapData|local_Random_TreemapData|global_Random_TreemapData;
+    private treemapData: TreemapNode;
     private branch: IBranchData[];
     private edgeList: EdgeInfo[];
     private totalLength;
@@ -39,7 +37,7 @@ export class EdgeMeasurement {
     private edgeCrossingCount: number;
     private totalBending;
 
-    constructor(treemapData: TreemapData|seqeunce_TreemapData|local_Random_TreemapData|global_Random_TreemapData, branch: IBranchData[]) {
+    constructor(treemapData: TreemapNode, branch: IBranchData[]) {
         this.treemapData = treemapData;
         this.branch = branch;
         this.edgeList = new Array<EdgeInfo>();
